\name{getInfo}
\alias{getInfo}
\title{Get Memory and Runtime Info}
\usage{
  getInfo()
}
\value{
  character vector of length 1 containing the memory and
  runtime info.
}
\description{
  Get Memory and Runtime Info, for example
}
\examples{
cat(getInfo())
}
\author{
  Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

