\name{getJavaInfo}
\alias{getJavaInfo}
\title{Get Memory and Runtime Info from JVM}
\usage{
  getJavaInfo()
}
\value{
  character vector of length 1 containing the memory and
  runtime info.
}
\description{
  Get Memory and Runtime Info from JVM
}
\examples{
cat(getJavaInfo())
}
\author{
  Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}

