% Generated by roxygen2 (4.0.1): do not edit by hand
\name{substituteEps}
\alias{substituteEps}
\title{Substitute Epsilon}
\usage{
substituteEps(graph, eps = 10^(-3))
}
\arguments{
\item{graph}{A graph of class \code{\link{graphMCP}} or class
\code{\link{entangledMCP}}.}

\item{eps}{A numeric scalar specifying a value for epsilon edges.}
}
\value{
A graph where all epsilons have been replaced with the given value.
}
\description{
Substitute Epsilon with a given value.
}
\details{
For details see the given references.
}
\examples{
graph <- improvedParallelGatekeeping()
graph
substituteEps(graph, eps=0.01)
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\seealso{
\code{\link{graphMCP}}, \code{\link{entangledMCP}}
}
\keyword{graphs}
\keyword{print}

