% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampSize.R
\name{sampSize}
\alias{sampSize}
\title{Sample size calculations}
\usage{
sampSize(powerReqFunc, target, effSize, endPoints, ...)
}
\arguments{
\item{powerReqFunc}{The power requirement function.}

\item{target}{For example \code{function(x) {any(x)}} for at least one rejection or more complex user defined functions like \code{function(x) {(x[1]&&x[3])||(x[2]&&x[4])}}.}

\item{effSize}{...}

\item{endPoints}{...}

\item{...}{...}
}
\value{
...
}
\description{
Sample size calculations
}
\examples{
f <- function(x){1/100*log(x)}
}

