% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampSize.R
\name{sampSize}
\alias{sampSize}
\title{Sample size calculations}
\usage{
sampSize(upperN, lowerN = floor(upperN/2), targFunc, target, tol = 0.001,
  alRatio, Ntype = c("arm", "total"), verbose = FALSE)

sampSize(upperN, lowerN = floor(upperN/2), targFunc, target, tol = 0.001,
  alRatio, Ntype = c("arm", "total"), verbose = FALSE)
}
\arguments{
\item{upperN}{\code{targFunc(upperN)} should be bigger than target (otherwise \code{upperN} is doubled until this is the case).}

\item{lowerN}{\code{targFunc(lowerN)} should be smaller than target (otherwise \code{lowerN} is halfed until this is the case).}

\item{targFunc}{The target (power) function that should be monotonically increasing in \code{n}.}

\item{target}{For example \code{function(x) {any(x)}} for at least one rejection or more complex user defined functions like \code{function(x) {(x[1]&&x[3])||(x[2]&&x[4])}}.}

\item{tol}{Tolerance: The function searches the \code{n} with \code{targFunc(n)-target<tol} and \code{targFunc(n)>target}.}

\item{alRatio}{Allocation ratio.}

\item{Ntype}{Either \code{"arm"} or \code{"total"}.}

\item{verbose}{Logical, whether verbose output should be printed.}

\item{powerReqFunc}{The power requirement function.}

\item{effSize}{...}

\item{endPoints}{...}

\item{target}{The target value. The function searches the \code{n} with \code{targFunc(n)-target<tol} and \code{targFunc(n)>target}.}
}
\value{
...

Integer value \code{n} (of type numeric) with \code{targFunc(n)-target<tol} and \code{targFunc(n)>target}.
}
\description{
Sample size calculations

Function for sample size calculation
}
\details{
For details see the manual and examples.
}
\examples{
f <- function(x){1/100*log(x)}
f <- function(x){1/100*log(x)}
gMCP:::sampSize(upperN=1000, targFunc=f, target=0.8, verbose=TRUE, alRatio=1)
gMCP:::sampSize(lowerN=1, upperN=1000, targFunc=f, target=0.8, verbose=TRUE, alRatio=1)
}

